% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{apply_flags}
\alias{apply_flags}
\title{Apply logic of flags in registers}
\usage{
apply_flags(instr, addr1, addr2, reg)
}
\arguments{
\item{instr}{asm instruction previously called}

\item{addr1}{first address argument to previously called instruction}

\item{addr2}{second address argument to previously called instruction}

\item{reg}{(global) registry environment}
}
\value{
either the address of the next instruction to call, or `NULL`
}
\description{
Apply logic of flags in registers
}
