% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin2ascii}
\alias{bin2ascii}
\title{Convert binary values to ASCII string}
\usage{
bin2ascii(bin)
}
\arguments{
\item{bin}{a vector of pairs of binary values (runs of 0s and 1s, separated by spaces).
Any linebreaks will be removed.}
}
\value{
the ASCII representation of the input as a single UTF8 string
}
\description{
Convert binary values to ASCII string
}
