% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{create_mem}
\alias{create_mem}
\title{Create a memory array}
\usage{
create_mem(len = 128)
}
\arguments{
\item{len}{length of array (bytes)}
}
\value{
a vector of length \code{len} filled with 0x00
}
\description{
\code{rx86} will assume this is available globally as \code{mem}.
}
