% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{create_reg}
\alias{create_reg}
\title{Create a set of registers}
\usage{
create_reg()
}
\value{
an \code{environment} containing zeroed registers
}
\description{
Creates a new environment with the following elements (set to 0x00). An
environment is used so that it is properly mutable.
}
\details{
eax = accumulator
ebx = base
ecx = counter
edx = data
esi = source
edi = destination
al = low bytes of accumulator
dl = low bytes of data
zf = zero flag
esp = stack pointer
}
\seealso{
https://web.archive.org/web/20191114093028/https://gerardnico.com/computer/cpu/register/general
}
