% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex2string}
\alias{hex2string}
\title{Convert hex values to ASCII string}
\usage{
hex2string(hex)
}
\arguments{
\item{hex}{a vector of pairs of hex values (potentially prefixed with "0x").

It is assumed that the string contains a null terminator (0x00) so this will
be retained. Any data following that will be discarded.}
}
\value{
the ASCII representation of the input as a single string
}
\description{
Convert hex values to ASCII string
}
