% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{insert_code}
\alias{insert_code}
\title{Inject code into memory}
\usage{
insert_code(code, memory)
}
\arguments{
\item{code}{parsed asm code with one line per \code{bytecode} hex value}

\item{memory}{memory array into which the code should be injected}
}
\value{
a copy of the \code{memory} object with bytes inserted at the relevant addresses.
}
\description{
Hex values from code will be inserted based on sanitized addresses.
Any remaining memory slots will be pre-allocated to 0x00
}
