% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcodes.R
\docType{data}
\name{opcodes}
\alias{opcodes}
\alias{call}
\alias{pop}
\alias{mov}
\alias{add}
\alias{sub}
\alias{xor}
\alias{cmp}
\alias{jmp}
\alias{jnz}
\alias{halt}
\title{opcodes (mnemonics) to separate from arguments}
\format{
An object of class \code{character} of length 13.
}
\usage{
opcodes

call(x)

pop(x)

mov(x, y)

add(x, y)

sub(x, y)

xor(x, y)

cmp(x, y)

jmp(x)

jnz(x)

halt(x)
}
\description{
opcodes (mnemonics) to separate from arguments

Opcodes (x86 assembly operations)
}
\keyword{internal}
