% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_parse.R
\name{parse}
\alias{parse}
\title{Parse asm code and seperate opcodes from arguments}
\usage{
parse(asm, opcodes = opcodes)
}
\arguments{
\item{asm}{a \code{\link[=data.frame]{data.frame()}} containing columns \code{addr} (addresses),
\code{bytecode} (bytecode), and \code{instr} (instructions).}

\item{opcodes}{list of opcodes to recognise}
}
\value{
the input \code{asm} with additional columns containing identified
information
}
\description{
Parse asm code and seperate opcodes from arguments
}
