% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{reg_or_val}
\alias{reg_or_val}
\title{Extract from registry or use direct val}
\usage{
reg_or_val(x)
}
\arguments{
\item{x}{address to lookup}
}
\value{
the value stored at address \code{x}
}
\description{
Extract from registry or use direct val
}
