% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{runasm}
\alias{runasm}
\title{Run asm code}
\usage{
runasm(asm, step = FALSE, prog = FALSE)
}
\arguments{
\item{asm}{x86 (disassembled) assembly code}

\item{step}{(logical) pause after every instruction?}

\item{prog}{(logical) display the memory (via \code{\link[=showmem]{showmem()}}) at every step?}
}
\value{
\code{invisible(NULL)} - called for the side-effect of modifying the global
\code{mem} and \code{registers} objects.
}
\description{
Parses code, inserts values into memory, then proceeds to evaluate
the instructions in order. Each operation is followed by a check of the
registers to help identify the next instruction - either directed by a
JMP or simply the next in order.
}
\details{
See \code{vignette("mult_code_petzold", package = "rx86")} for an example of usage.
}
