% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize memory locations to strings}
\usage{
sanitize(x)
}
\arguments{
\item{x}{address to be sanitized}
}
\value{
a character string which can be used to lookup memory locations
}
\description{
Because memory is addressed by the names of the vectors (not positions) but
\emph{can} be addressed by e.g. 0x00 (which won't exist) we need to sanitize
addressed before using them to access memory. This is achieved by converting
to character with a "0x" prefix.
}
