% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{showmem}
\alias{showmem}
\title{Clean print method for memory and registers}
\usage{
showmem()
}
\value{
\code{invisible(NULL)} - called for the side-effect of printing to console
}
\description{
Clean print method for memory and registers
}
