% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_parse.R
\name{split_args}
\alias{split_args}
\title{Split argument list into individual elements}
\usage{
split_args(args)
}
\arguments{
\item{args}{string containing 1, or 2 comma-separated arguments}
}
\value{
a [data.frame()] containing columns `arg1` and `arg2`
}
\description{
Split argument list into individual elements
}
