% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_parse.R
\name{split_pairs}
\alias{split_pairs}
\title{Split byte strings into pairs of numbers}
\usage{
split_pairs(x, split = "")
}
\arguments{
\item{x}{string of bytes}

\item{split}{character at which to \code{\link[=strsplit]{strsplit()}}}
}
\value{
a string of paired bytes, comma-separated
}
\description{
Split byte strings into pairs of numbers
}
